refactor(mcp): decouple pkg/runtime and pkg/agent from the MCP toolset package
pkg/runtime and pkg/agent imported pkg/tools/mcp for a handful of small
helpers, linking the entire MCP toolset - and through it expr-lang/expr
(via toolinstall) and goja (via pkg/js) - into every embedder, even ones
with zero MCP toolsets.

- Move the interactive-prompt gating helpers, OAuth sentinel errors and
  PromptInfo/PromptArgument to pkg/tools; pkg/tools/mcp keeps aliases
  and forwarders so its public API is unchanged.
- Move the OAuth authorization-code flow (PKCE/state, callback server,
  token exchange/refresh, dynamic client registration) to the new leaf
  pkg/tools/mcp/oauthflow, again with aliases and forwarders; only the
  upstream-header-scoping transport stays in pkg/tools/mcp.
- Runtime prompt discovery now goes through a small mcpPromptToolset
  interface (satisfied by *mcp.Toolset) instead of a concrete type
  assertion.
