+++ pkg/agent/agent.go
@@ -16,7 +16,6 @@ import (
 	"github.com/docker/docker-agent/pkg/config/types"
 	"github.com/docker/docker-agent/pkg/model/provider"
 	"github.com/docker/docker-agent/pkg/tools"
-	mcptools "github.com/docker/docker-agent/pkg/tools/mcp"
 )
 
 // Agent represents an AI agent
@@ -570,7 +569,7 @@ func (a *Agent) ensureToolSetsAreStarted(ctx context.Context) {
 			continue
 		}
 		desc := tools.DescribeToolSet(toolSet)
-		if mcptools.IsAuthorizationRequired(err) {
+		if tools.IsAuthorizationRequired(err) {
 			// Recovery: previously-working toolset lost its OAuth token in the
 			// background. Emit the targeted re-auth notice once per streak so the
 			// user knows a dialog will appear on their next message.
